refactor(agent): give every service a file named after it
The seventeen services were scattered across thirteen domain folders,
three of them in a file called manager.ts. They now sit in services/, one
per file, named for what they export, as apps/api already does.

A .service.ts file holds the service and its layer and nothing else, so
the modules that mixed the two are split: the artifact store leaves its
verification behind, the control plane leaves its client, the agent state
leaves its transitions, and the command runner leaves the vocabulary its
callers speak.
