+++ apps/agent/src/exports/manager.ts
@@ -1,17 +1,4 @@
-import { FileSystem, Path } from '@effect/platform';
-import type { DesiredArtifact, DesiredExport, ReportedExport } from '@repo/protocol';
-import { Effect } from 'effect';
-import { s3Credentials } from '#aws/credentials.ts';
-import { AwsCredentialProvider } from '#aws/provider.ts';
-import { AgentConfig } from '#config.ts';
-import { writeBundle } from '#exports/bundle.ts';
-import { nowTimestamp } from '#lib/clock.ts';
-import { ZerofsTopology } from '#volumes/topology.ts';
-import { flush } from '#volumes/zerofs.ts';
-
-const UPLOAD_PART_SIZE_BYTES = 8_388_608;
-const UPLOAD_QUEUE_SIZE = 4;
-const UPLOAD_RETRIES = 3;
+import type { ReportedExport } from '@repo/protocol';
 
 const isObject = (value: unknown): value is Record<string, unknown> =>
   value !== null && typeof value === 'object' && !Array.isArray(value);
@@ -23,89 +10,3 @@ const isReportedExport = (value: unknown): value is ReportedExport =>
 export function readExportReports(value: unknown): ReportedExport[] {
   return Array.isArray(value) ? value.filter(isReportedExport) : [];
 }
-
-export class ExportManager extends Effect.Service<ExportManager>()('ExportManager', {
-  effect: Effect.gen(function* () {
-    const config = yield* AgentConfig;
-    const fs = yield* FileSystem.FileSystem;
-    const path = yield* Path.Path;
-    const topology = yield* ZerofsTopology;
-    const credentials = yield* AwsCredentialProvider;
-
-    /**
-     * Streamed in parts rather than read into memory: a bundle is a tenant's whole dataset. A
-     * failure before `end()` leaves the multipart upload uncommitted rather than publishing a
-     * truncated bundle, and the bucket's abort rule reaps it — which is why exports need no
-     * delete permission.
-     */
-    const upload = Effect.fn('ExportManager.upload')(function* ({
-      bundlePath,
-      objectKey,
-    }: {
-      bundlePath: string;
-      objectKey: string;
-    }) {
-      yield* Effect.annotateCurrentSpan({ objectKey });
-      const resolved = yield* credentials.resolve;
-      const client = new Bun.S3Client({
-        bucket: config.exportBucket,
-        region: config.awsRegion,
-        ...s3Credentials(resolved),
-      });
-      yield* Effect.tryPromise(async () => {
-        const writer = client.file(objectKey).writer({
-          partSize: UPLOAD_PART_SIZE_BYTES,
-          queueSize: UPLOAD_QUEUE_SIZE,
-          retry: UPLOAD_RETRIES,
-        });
-        for await (const chunk of Bun.file(bundlePath).stream()) {
-          writer.write(chunk);
-        }
-        await writer.end();
-      });
-    });
-
-    /**
-     * The staging tree is a second copy of a tenant's dataset in the clear on a shared host, so
-     * it is removed whether or not the upload worked.
-     */
-    const write = Effect.fn('ExportManager.write')(function* ({
-      desired,
-      artifact,
-      devicePath,
-    }: {
-      desired: DesiredExport;
-      artifact: DesiredArtifact;
-      devicePath: string;
-    }) {
-      yield* Effect.annotateCurrentSpan({ exportId: desired.exportId });
-      const stagingDir = path.join(config.exportStagingDir, desired.exportId);
-      return yield* Effect.ensuring(
-        Effect.gen(function* () {
-          // Under `ignore_fsync` the guest's barriers are not durability points, so this is what
-          // turns its acknowledged writes into bytes the device will hand back.
-          yield* flush(topology.place().admin);
-          const bundle = yield* writeBundle({ artifact, devicePath, stagingDir });
-          yield* upload({ bundlePath: bundle.path, objectKey: desired.objectKey });
-          yield* Effect.logInfo('export written').pipe(
-            Effect.annotateLogs({
-              exportId: desired.exportId,
-              objectKey: desired.objectKey,
-              sizeBytes: bundle.sizeBytes,
-            }),
-          );
-          return {
-            exportId: desired.exportId,
-            state: 'ready',
-            sizeBytes: bundle.sizeBytes,
-            readyAt: yield* nowTimestamp,
-          } satisfies ReportedExport;
-        }),
-        fs.remove(stagingDir, { recursive: true, force: true }).pipe(Effect.ignore),
-      );
-    });
-
-    return { write };
-  }),
-  dependencies: [AgentConfig.Default, ZerofsTopology.Default, AwsCredentialProvider.Default],
-}) {}
