+++ apps/agent/src/index.ts
@@ -1,24 +1,24 @@
 import { FetchHttpClient } from '@effect/platform';
 import { BunContext, BunRuntime } from '@effect/platform-bun';
 import { Effect, Layer } from 'effect';
-import { DesiredStateCache } from '#agent/desired-state.ts';
 import { run } from '#agent/run.ts';
-import { AgentSessionHolder } from '#agent/session.ts';
-import { AgentConfig } from '#config.ts';
-import * as ControlPlane from '#control/client.ts';
-import { ExportManager } from '#exports/manager.ts';
-import * as Exec from '#lib/exec.ts';
 import { AgentLogger } from '#lib/logger.ts';
-import { TenantLogQueue } from '#logs/queue.ts';
-import { TenantLogReceiver } from '#logs/receiver.ts';
-import { SlotAllocator } from '#network/allocator.ts';
-import { CaddyProxy } from '#proxy/caddy.ts';
-import { Reconciler } from '#reconcile/reconciler.ts';
-import * as AgentState from '#reconcile/state.ts';
-import * as Artifacts from '#vm/artifacts.ts';
-import { VmManager } from '#vm/manager.ts';
-import { VolumeManager } from '#volumes/manager.ts';
-import { ZerofsTopology } from '#volumes/topology.ts';
+import { AgentConfig } from '#services/agent-config.service.ts';
+import { AgentSessionHolder } from '#services/agent-session-holder.service.ts';
+import * as AgentState from '#services/agent-state.service.ts';
+import * as Artifacts from '#services/artifact-store.service.ts';
+import { CaddyProxy } from '#services/caddy-proxy.service.ts';
+import * as Exec from '#services/command-runner.service.ts';
+import * as ControlPlane from '#services/control-plane.service.ts';
+import { DesiredStateCache } from '#services/desired-state-cache.service.ts';
+import { ExportManager } from '#services/export-manager.service.ts';
+import { Reconciler } from '#services/reconciler.service.ts';
+import { SlotAllocator } from '#services/slot-allocator.service.ts';
+import { TenantLogQueue } from '#services/tenant-log-queue.service.ts';
+import { TenantLogReceiver } from '#services/tenant-log-receiver.service.ts';
+import { VmManager } from '#services/vm-manager.service.ts';
+import { VolumeManager } from '#services/volume-manager.service.ts';
+import { ZerofsTopology } from '#services/zerofs-topology.service.ts';
 
 const platform = Layer.mergeAll(BunContext.layer, FetchHttpClient.layer);
