+++ apps/agent/src/network/nftables.ts
@@ -1,5 +1,5 @@
-import { stdoutOf } from '#lib/exec.ts';
 import { type FirewallState, renderRuleset } from '#network/firewall.ts';
+import { stdoutOf } from '#services/command-runner.service.ts';
 
 export const applyRuleset = (state: FirewallState) =>
   stdoutOf({ command: ['nft', '-f', '-'], stdin: renderRuleset(state) });
