+++ apps/agent/src/services/aws-credential-provider.service.ts
