+++ apps/agent/tests/services/tenant-log-receiver.test.ts
@@ -3,8 +3,8 @@ import { Buffer } from 'node:buffer';
 import { createConnection, type Socket } from 'node:net';
 import { join } from 'node:path';
 import { Effect, Layer } from 'effect';
-import { TenantLogQueue } from '#logs/queue.ts';
-import { TenantLogReceiver } from '#logs/receiver.ts';
+import { TenantLogQueue } from '#services/tenant-log-queue.service.ts';
+import { TenantLogReceiver } from '#services/tenant-log-receiver.service.ts';
 import { LOG_SOURCE } from '#tests/support/fixtures.ts';
 import { gapFrame, guestLogFrame } from '#tests/support/guest-frames.ts';
 import { drainedEvents } from '#tests/support/logs.ts';
